update xivo-migrate-db-91-to-94
First, don't call apt-get from the script, now that it is running in the
postinst of a Debian package.

Don't save the postgresql configuration file, since they aren't lost
during the cluster upgrade: they are just used at the configuration for
the new cluster.

Remove extraneous check on ports: pg_upgradecluster already does this
kind of check, i.e. let pg_upgradecluster do its own checks, which are
better than ours anyway.

Better handling of error during the upgrade. Previous code was not even
detecting properly when pg_upgradecluster was failing (because it was
checking the status code of a pipe commande and the shell pipefail
option wasn't set, and the exit status was always 0).

Only log the output of pg_upgradecluster in the migration file, the rest
of the information is already available in the other log files, and the
output of pg_upgradecluster is mostly noise as far as we are concerned.