update postgresql-9.4 dependency
This is a temporary workaround for the following problems:

* a newer version of postgresql-9.4 is available on the Debian mirrors
* a newer version of xivo-dird is available on the XiVO mirror
* someone run xivo-upgrade

What can happens then is that during the xivo-upgrade:

* postgresql-9.4 is unpacked (and the cluster is then stopped)
* xivo-dird is unpackaged then configured, but it doesn't try to
  update it's database schema since the postgresql cluster is not
  running at this stage
* postgresql-9.4 is configured

So you find yourself at the end of the xivo-upgrade with an not-updated
database schema, and this is not good.

We're still looking on the right way to fix it, but right now, adding a
dependecy on the latest postgresql-9.4 version (and updating the
xivo-dbms dependecy in xivo-dird and xivo-manage-db) will fix the issue,
until Debian release another version of the postgresql-9.4 package.